patch 1.0.0

+++ Services/NotifyTimerService.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Timers;
 using TamagotchiBot.Database;
 using TamagotchiBot.Services.Mongo;
@@ -16,6 +17,7 @@ namespace TamagotchiBot.Services
     {
         private Timer _notifyTimer;
         private Timer _changelogsTimer;
+        private Timer _dailyRewardTimer;
 
         private readonly ITelegramBotClient _botClient;
         private PetService _petService;
@@ -79,23 +81,32 @@ namespace TamagotchiBot.Services
             _changelogsTimer.AutoReset = false;
             _changelogsTimer.Enabled = true;
         }
-        private async void OnChangelogsTimedEvent(object sender, ElapsedEventArgs e)
+
+        public void SetDailyRewardNotificationTimer()
         {
-            var usersToNotify = GetAllActiveUsersIds();
-            Log.Information($"Changelog timer - {usersToNotify.Count} users");
-            _sinfoService.DisableChangelogsSending();
+            TimeSpan timeToWait = TimeSpan.FromMinutes(3);
+            Log.Information("DailyRewardNotification timer set to wait for " + timeToWait.TotalSeconds + "s");
+            _dailyRewardTimer = new Timer(timeToWait);
+            _dailyRewardTimer.Elapsed += OnDailyRewardTimedEvent;
+            _dailyRewardTimer.AutoReset = true;
+            _dailyRewardTimer.Enabled = true;
+        }
+
+        private async void OnDailyRewardTimedEvent(object sender, ElapsedEventArgs e)
+        {
+            var usersToNotify = GetAllDailyRewardUsersIds();
+            Log.Information($"DailyRewardNotification timer - {usersToNotify.Count} users");
             foreach (var userId in usersToNotify)
             {
-                var user = _userService.Get(long.Parse(userId));
+                var user = _userService.Get(userId);
                 Resources.Resources.Culture = new CultureInfo(user?.Culture ?? "ru");
 
                 try
                 {
-                    await _botClient.SendStickerAsync(userId, Constants.StickersId.ChangelogSticker);
-                    await _botClient.SendTextMessageAsync(userId, Resources.Resources.changelog1Text);
+                    await _botClient.SendStickerAsync(userId, GetRandomDailyRewardSticker());
+                    await _botClient.SendTextMessageAsync(userId, Resources.Resources.rewardNotification);
 
-
-                    Log.Information($"Sent changelog to '@{user.Username}'");
+                    Log.Information($"Sent DailyRewardNotification to '@{user.Username}'");
                 }
                 catch (ApiRequestException ex)
                 {
@@ -114,7 +125,63 @@ namespace TamagotchiBot.Services
                     Log.Error(ex.Message);
                 }
             }
+        }
+        private async void OnChangelogsTimedEvent(object sender, ElapsedEventArgs e)
+        {
+            var usersToNotify = GetAllUsersIds();
+            Log.Information($"Changelog timer - {usersToNotify.Count} users");
+            _sinfoService.DisableChangelogsSending();
 
+            int usersSuccess = 0;
+            foreach (var userId in usersToNotify)
+            {
+                var user = _userService.Get(long.Parse(userId));
+                Resources.Resources.Culture = new CultureInfo(user?.Culture ?? "ru");
+
+                try
+                {
+                    await _botClient.SendStickerAsync(userId, Constants.StickersId.ChangelogSticker);
+                    await _botClient.SendTextMessageAsync(userId, Resources.Resources.changelog1Text);
+                    await Task.Delay(1000);
+
+                    usersSuccess++;
+                    Log.Information($"Sent changelog to '@{user.Username}'");
+                }
+                catch (ApiRequestException ex)
+                {
+                    if (ex == null)
+                    {
+                        Log.Error("HZ: ApiRequestException is null 0_o");
+                        continue;
+                    }
+
+                    if (ex.ErrorCode == 403) //Forbidden by user
+                    {
+                        if (ex == null)
+                        {
+                            Log.Error("HZ: ApiRequestException is null 0_o");
+                            continue;
+                        }
+
+                        Log.Warning($"{ex?.Message} @{user?.Username}, id: {user?.UserId}");
+
+                        //remove all data about user
+
+                        if (user == null)
+                            continue;
+
+                        _chatService.Remove(user.UserId);
+                        _petService.Remove(user.UserId);
+                        _userService.Remove(user.UserId);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    Log.Error(ex.Message);
+                }
+            }
+
+            Log.Information($"Changelogs have been sent - {usersSuccess} success, {usersToNotify.Count - usersSuccess} failed");
         }
         private async void OnNotifyTimedEvent(object sender, ElapsedEventArgs e)
         {
@@ -138,48 +205,59 @@ namespace TamagotchiBot.Services
             _sinfoService.UpdateNextNotify(_nextNotify);
 
             var usersToNotify = GetUserIdToNotify();
-            Log.Information($"Notify timer - {usersToNotify.Count} users");
-            foreach (var userId in usersToNotify)
+            if (usersToNotify != null)
             {
-                var user = _userService.Get(long.Parse(userId));
-                Resources.Resources.Culture = new CultureInfo(user?.Culture ?? "ru");
-                int rand = new Random().Next(3);
-                var notifyText = new List<string>()
+                Log.Information($"Notify timer - {usersToNotify.Count} users");
+                foreach (var userId in usersToNotify)
+                {
+                    var user = _userService.Get(long.Parse(userId));
+                    Resources.Resources.Culture = new CultureInfo(user?.Culture ?? "ru");
+                    int rand = new Random().Next(3);
+                    var notifyText = new List<string>()
                 {
                     Resources.Resources.ReminderNotifyText1,
                     Resources.Resources.ReminderNotifyText2,
                     Resources.Resources.ReminderNotifyText3
                 };
 
-                string toSendText = notifyText.ElementAtOrDefault(rand) ?? Resources.Resources.ReminderNotifyText1;
-                try
-                {
-                    await _botClient.SendStickerAsync(userId, Constants.StickersId.PetBored_Cat);
-                    await _botClient.SendTextMessageAsync(userId, toSendText);
-
-                    Log.Information($"Sent reminder to '@{user?.Username ?? userId}'");
-                }
-                catch (ApiRequestException ex)
-                {
-                    if (ex.ErrorCode == 403) //Forbidden by user
+                    string toSendText = notifyText.ElementAtOrDefault(rand) ?? Resources.Resources.ReminderNotifyText1;
+                    try
                     {
-                        Log.Warning($"{ex.Message} @{user?.Username ?? userId}, id: {userId}");
+                        await Task.Delay(5000); //pause between notification to avoid 429 error
+                        await _botClient.SendStickerAsync(userId, Constants.StickersId.PetBored_Cat);
+                        await _botClient.SendTextMessageAsync(userId, toSendText);
 
-                        _chatService.Remove(user.UserId);
-                        _petService.Remove(user.UserId);
-                        _userService.Remove(user.UserId);
+                        Log.Information($"Sent reminder to '@{user?.Username ?? userId}'");
+                    }
+                    catch (ApiRequestException ex)
+                    {
+                        if (ex.ErrorCode == 403) //Forbidden by user
+                        {
+                            Log.Warning($"{ex.Message} @{user?.Username ?? userId}, id: {userId}");
+
+                            _chatService.Remove(user.UserId);
+                            _petService.Remove(user.UserId);
+                            _userService.Remove(user.UserId);
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        Log.Error(ex.Message);
                     }
                 }
-                catch (Exception ex)
-                {
-                    Log.Error(ex.Message);
-                }
             }
-            Log.Information("Notifications completed!");      
+
+            Log.Information("Notifications completed!");
             Log.Information($"Next notification: {_nextNotify} UTC || remaining {_notifyEvery:c}");
         }
         private async void SendDevNotify()
         {
+            if (_envs?.ChatsToDevNotify == null)
+            {
+                Log.Warning("No chats do DEV notify");
+                return;
+            }
+
             var chatsToNotify = new List<string>(_envs.ChatsToDevNotify);
 
             foreach (var chatId in chatsToNotify)
@@ -190,7 +268,7 @@ namespace TamagotchiBot.Services
 
                     var dailyInfoToday = _dailyInfoService.GetToday();
 
-                    if (dailyInfoToday == null || 
+                    if (dailyInfoToday == null ||
                         (dailyInfoToday != null && (DateTime.UtcNow - dailyInfoToday.DateInfo) > _envs.DevExtraNotifyEvery))
                     {
                         Log.Information("Sent extra dev notify");
@@ -251,10 +329,33 @@ namespace TamagotchiBot.Services
                     usersToNotify.Add(pet.UserId.ToString());
             }
 
+            if (_envs.AlwaysNotifyUsers == null)
+            {
+                Log.Warning("No always notify persons");
+                return null;
+            }
+
             foreach (var userAlwaysNotify in _envs.AlwaysNotifyUsers)
                 if (!usersToNotify.Contains(userAlwaysNotify))
                     usersToNotify.Add(userAlwaysNotify);
 
+            return usersToNotify;
+        }
+        private List<long> GetAllDailyRewardUsersIds()
+        {
+            List<long> usersToNotify = new();
+            var petsDB = _petService.GetAll();
+
+            foreach (var pet in petsDB)
+            {
+                var user = _userService.Get(pet.UserId);
+
+                if (user.NextDailyRewardNotificationTime < DateTime.UtcNow && pet.GotDailyRewardTime.AddDays(1) < DateTime.UtcNow)
+                {
+                    _userService.UpdateNextDailyRewardNotificationTime(user.UserId, DateTime.UtcNow.AddDays(1));
+                    usersToNotify.Add(user.UserId);
+                }
+            }
             return usersToNotify;
         }
         private List<string> GetAllActiveUsersIds()
@@ -266,6 +367,31 @@ namespace TamagotchiBot.Services
                 usersToNotify.Add(pet.UserId.ToString());
 
             return usersToNotify;
+        }
+        private List<string> GetAllUsersIds()
+        {
+            List<string> usersToNotify = new();
+            var usersAUDS = _allUsersDataService.GetAll();
+
+            foreach (var user in usersAUDS)
+                usersToNotify.Add(user.UserId.ToString());
+
+            return usersToNotify;
+        }
+
+        private string GetRandomDailyRewardSticker()
+        {
+            var random = new Random().Next(0, 6);
+
+            return random switch
+            {
+                1 => Constants.StickersId.DailyRewardNotificationSticker_1,
+                2 => Constants.StickersId.DailyRewardNotificationSticker_2,
+                3 => Constants.StickersId.DailyRewardNotificationSticker_3,
+                4 => Constants.StickersId.DailyRewardNotificationSticker_4,
+                5 => Constants.StickersId.DailyRewardNotificationSticker_5,
+                _ => Constants.StickersId.DailyRewardNotificationSticker_3,
+            };
         }
     }
 }
